refactor(protocol): extract decryption queue into own module with private fields
Move PendingDecryptionQueue and related types from types.rs into
decryption_queue.rs, shorten redundant field names (e.g.
pending_decryption → queues), and make all fields private behind
accessor methods to reduce test coupling to internal layout.
